+++ src/main/java/backend/wal/reservation/adapter/ReservationSchedulerAdapter.java
@@ -1,6 +1,6 @@
 package backend.wal.reservation.adapter;
 
-import backend.wal.reservation.application.port.ReservationSchedulerPort;
+import backend.wal.reservation.application.port.out.ReservationSchedulerPort;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
