+++ src/main/java/backend/wal/reservation/domain/Reservations.java
@@ -1,7 +1,7 @@
 package backend.wal.reservation.domain;
 
-import backend.wal.reservation.application.port.dto.ReservationCalendarResponseDto;
-import backend.wal.reservation.application.port.dto.ReservationHistoryResponseDto;
+import backend.wal.reservation.application.port.in.dto.ReservationCalendarResponseDto;
+import backend.wal.reservation.application.port.in.dto.ReservationHistoryResponseDto;
 import backend.wal.reservation.domain.aggregate.Reservation;
 
 import java.util.List;
