+++ src/main/java/backend/wal/reservation/domain/aggregate/Reservation.java
@@ -1,7 +1,7 @@
 package backend.wal.reservation.domain.aggregate;
 
-import backend.wal.reservation.application.port.dto.ReservationHistoryResponseDto;
-import backend.wal.reservation.application.port.dto.AddReservationRequestDto;
+import backend.wal.reservation.application.port.in.dto.ReservationHistoryResponseDto;
+import backend.wal.reservation.application.port.in.dto.AddReservationRequestDto;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
