+++ src/main/java/backend/wal/reservation/web/dto/AddReservationRequest.java
@@ -1,7 +1,7 @@
 package backend.wal.reservation.web.dto;
 
 import backend.wal.reservation.domain.aggregate.ShowStatus;
-import backend.wal.reservation.application.port.dto.AddReservationRequestDto;
+import backend.wal.reservation.application.port.in.dto.AddReservationRequestDto;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
