+++ src/main/java/backend/wal/reservation/web/dto/ReservationHistoryResponse.java
@@ -1,6 +1,6 @@
 package backend.wal.reservation.web.dto;
 
-import backend.wal.reservation.application.port.dto.ReservationHistoryResponseDto;
+import backend.wal.reservation.application.port.in.dto.ReservationHistoryResponseDto;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
