+++ src/test/java/backend/wal/reservation/domain/aggregate/entity/ReservationTest.java
@@ -1,6 +1,6 @@
 package backend.wal.reservation.domain.aggregate.entity;
 
-import backend.wal.reservation.application.port.dto.AddReservationRequestDto;
+import backend.wal.reservation.application.port.in.dto.AddReservationRequestDto;
 import backend.wal.reservation.domain.aggregate.Reservation;
 import backend.wal.reservation.domain.aggregate.ShowStatus;
 import org.junit.jupiter.api.DisplayName;
